fix PGH003
