+++ scripts/tools/filter_logs.py
@@ -1,14 +1,14 @@
 # /// script
 # requires-python = ">=3.11"
 # dependencies = [
-#     "alive-progress",
+#     "rich",
 # ]
 # ///
 import argparse
 from collections.abc import Iterable
 from pathlib import Path
 
-from alive_progress import alive_it  # type: ignore
+from rich.progress import track
 
 LEVELS_TO_INCLUDE = {"WARNING", "ERROR", "CRITICAL"}
 
@@ -19,7 +19,8 @@ def filter_log_file(log_file: Path) -> Iterable[str]:
     print(f"Filtering: {log_file.resolve()}")
     log_content = log_file.read_text(encoding="utf8")
     last_level = None
-    for line in alive_it(log_content.splitlines()):
+    lines = log_content.splitlines()
+    for line in track(lines, total=len(lines)):
         level = line[LEVEL_AT : LEVEL_AT + 9].strip()
         if level:
             last_level = level
